+++ system/modules/calendar/classes/Events.php
@@ -451,6 +451,16 @@ abstract class Events extends \Module
 				return array($objToday->dayBegin, $objToday->yearEnd, $GLOBALS['TL_LANG']['MSC']['cal_empty']);
 				break;
 
+			case 'next_next_month':
+				$objToday = new \Date();
+				return array(($objToday->monthEnd + 1), strtotime('+1 month', $objToday->monthEnd), $GLOBALS['TL_LANG']['MSC']['cal_empty']);
+				break;
+
+			case 'next_next_year':
+				$objToday = new \Date();
+				return array(($objToday->yearEnd + 1), strtotime('+1 year', $objToday->yearEnd), $GLOBALS['TL_LANG']['MSC']['cal_empty']);
+				break;
+
 			case 'next_all': // 2038-01-01 00:00:00
 				$objToday = new \Date();
 				return array($objToday->dayBegin, 2145913200, $GLOBALS['TL_LANG']['MSC']['cal_empty']);
@@ -499,6 +509,16 @@ abstract class Events extends \Module
 			case 'past_cur_year':
 				$objToday = new \Date();
 				return array($objToday->yearBegin, ($objToday->dayBegin - 1), $GLOBALS['TL_LANG']['MSC']['cal_empty']);
+				break;
+
+			case 'past_prev_month':
+				$objToday = new \Date();
+				return array(strtotime('-1 month', $objToday->monthBegin), ($objToday->monthBegin - 1), $GLOBALS['TL_LANG']['MSC']['cal_empty']);
+				break;
+
+			case 'past_prev_year':
+				$objToday = new \Date();
+				return array(strtotime('-1 year', $objToday->yearBegin), ($objToday->yearBegin - 1), $GLOBALS['TL_LANG']['MSC']['cal_empty']);
 				break;
 
 			case 'past_all': // 1970-01-01 00:00:00